requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
aws-sam-translator==1.34.0
aws-xray-sdk==2.6.0
boto==2.49.0
boto3==1.9.250
botocore==1.12.250
certifi==2020.12.5
cffi==1.14.4
cfn-lint==0.44.6
chardet==3.0.4
cryptography==3.3.1
Cython==0.29.21
decorator==4.4.2
docker==4.4.1
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
idna==2.8
iniconfig==1.1.1
Jinja2==2.11.3
jmespath==0.10.0
jsondiff==1.2.0
jsonpatch==1.28
jsonpickle==1.5.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
MarkupSafe==1.1.1
mock==4.0.3
more-itertools==8.6.0
moto==1.3.16
networkx==2.5
numpy==1.20.0
oauthlib==3.1.0
packaging==20.9
pandas==1.2.1
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.2.2
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2021.1
PyYAML==5.1.2
requests==2.22.0
requests-mock==1.7.0
requests-oauthlib==1.2.0
responses==0.12.1
rsa==4.7
s3transfer==0.2.1
six==1.15.0
sshpubkeys==3.3.1
toml==0.10.2
urllib3==1.25.11
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.0
